api/requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-Cors==5.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
numpy==2.2.2
opencv-python-headless==4.11.0.86
packaging==24.2
pillow==11.1.0
python-dotenv==1.0.1
Werkzeug==3.1.3
easyocr==1.7.1
PyJWT==2.8.0